+++ app/src/main/java/com/spaceapp/core/designsystem/components/DefaultUnderline.kt
@@ -1,4 +1,4 @@
-package com.spaceapp.core.designsystem.component
+package com.spaceapp.core.designsystem.components
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.padding
