+++ app/src/main/java/com/spaceapp/presentation/forgot_password/components/SendPasswordResetMail.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.spaceapp.R
 import com.spaceapp.core.common.helper.MobileServiceType
-import com.spaceapp.core.designsystem.component.DefaultButton
+import com.spaceapp.core.designsystem.components.DefaultButton
 import com.spaceapp.core.designsystem.theme.White
 import com.spaceapp.presentation.utils.ForgotPasswordScreenConstants
