+++ app/src/main/java/com/spaceapp/presentation/universe_glossary/UniverseGlossaryScreen.kt
@@ -2,6 +2,7 @@ package com.spaceapp.presentation.universe_glossary
 
 import android.app.Activity
 import androidx.activity.compose.BackHandler
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -10,14 +11,18 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.spaceapp.core.designsystem.component.ErrorCard
-import com.spaceapp.core.designsystem.component.LoadingSpinner
+import com.spaceapp.R
+import com.spaceapp.core.designsystem.components.DefaultOutlinedTextField
+import com.spaceapp.core.designsystem.components.ErrorCard
+import com.spaceapp.core.designsystem.components.LoadingSpinner
 import com.spaceapp.domain.model.glossary_data.GlossaryContent
-import com.spaceapp.presentation.universe_glossary.components.SearchField
-import com.spaceapp.presentation.universe_glossary.components.SearchResultEmpty
 import com.spaceapp.presentation.universe_glossary.components.StatefulGlossaryCard
 import com.spaceapp.presentation.universe_glossary.state.GlossaryState
 import com.spaceapp.presentation.utils.GlossaryImageType
@@ -64,12 +69,15 @@ private fun UniverseGlossaryContent(
             text = constants.title,
             style = MaterialTheme.typography.headlineLarge
         )
-        SearchField(
+        DefaultOutlinedTextField(
             modifier = modifier
                 .fillMaxWidth()
                 .padding(top = 16.dp, start = 16.dp, end = 16.dp),
-            searchValue = searchValue,
-            onSearchValChange = onSearchValChange
+            onValueChanged = onSearchValChange,
+            labelText = UniverseGlossaryScreenConstants.search_field_text,
+            keyboardType = KeyboardType.Text,
+            leadingIconId = R.drawable.ic_baseline_search,
+            value = searchValue
         )
         when (glossaryState) {
             is GlossaryState.Loading -> {
@@ -99,7 +107,26 @@ private fun UniverseGlossaryContent(
 @Composable
 private fun TermList(filteredGlossaryList: List<GlossaryContent>) {
     if (filteredGlossaryList.isEmpty()) {
-        SearchResultEmpty()
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(horizontal = 16.dp),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            Image(
+                modifier = Modifier.size(72.dp),
+                painter = painterResource(id = R.drawable.empty_list),
+                contentDescription = null,
+                contentScale = ContentScale.Fit
+            )
+            Text(
+                modifier = Modifier.padding(top = 16.dp),
+                text = UniverseGlossaryScreenConstants.no_result,
+                style = MaterialTheme.typography.headlineSmall,
+                textAlign = TextAlign.Center
+            )
+        }
     } else {
         LazyColumn(
             modifier = Modifier
