--- app/src/main/java/com/spaceapp/presentation/universe_glossary/components/SearchResultEmptyView.kt
@@ -1,43 +1,0 @@
-package com.spaceapp.presentation.universe_glossary.components
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import com.spaceapp.R
-import com.spaceapp.presentation.utils.UniverseGlossaryScreenConstants
-
-@Composable
-fun SearchResultEmpty() {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(horizontal = 16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Image(
-            modifier = Modifier.size(72.dp),
-            painter = painterResource(id = R.drawable.empty_list),
-            contentDescription = null,
-            contentScale = ContentScale.Fit
-        )
-        Text(
-            modifier = Modifier.padding(top = 16.dp),
-            text = UniverseGlossaryScreenConstants.no_result,
-            style = MaterialTheme.typography.headlineSmall,
-            textAlign = TextAlign.Center
-        )
-    }
-}
